requirements.txt
# albumentations
numpy
nibabel
matplotlib
pyyaml
tqdm
torch
torchvision
torchaudio
torchsummary
